separate out logic to set config defaults
the use case is i want to try and setup PeachCloud services to follow Linux best practices with regards to where config (and other) files are located, so i want `ssb-config` without `rc`.
