+++ test.js
@@ -1,3 +1,2 @@
-
-
 console.log(require('./inject')('testnet', {port: 9999, friends: {dunbar: 1500}}))
+console.log(require('./inject')('ssb'))
